+++ src/test/java/de/adorsys/keycloak/config/service/ImportClientsIT.java
@@ -26,12 +26,16 @@ import de.adorsys.keycloak.config.exception.ImportProcessingException;
 import de.adorsys.keycloak.config.exception.KeycloakRepositoryException;
 import de.adorsys.keycloak.config.model.RealmImport;
 import de.adorsys.keycloak.config.properties.ImportConfigProperties;
+import de.adorsys.keycloak.config.properties.KeycloakConfigProperties;
 import de.adorsys.keycloak.config.util.VersionUtil;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
+import org.keycloak.authorization.client.AuthzClient;
+import org.keycloak.authorization.client.Configuration;
 import org.keycloak.representations.idm.*;
 import org.keycloak.representations.idm.authorization.*;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -50,6 +54,9 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 class ImportClientsIT extends AbstractImportTest {
     private static final String REALM_NAME = "realmWithClients";
     private static final String REALM_AUTH_FLOW_NAME = "realmWithClientsForAuthFlowOverrides";
+
+    @Autowired
+    private KeycloakConfigProperties properties;
 
     ImportClientsIT() {
         this.resourcePath = "import-files/clients";
@@ -525,16 +532,16 @@ class ImportClientsIT extends AbstractImportTest {
 
         RealmImport foundImport3 = getFirstImport("10.3_update_realm__raise_error_update_authorization_client_bearer_only.json");
         thrown = assertThrows(ImportProcessingException.class, () -> realmImportService.doImport(foundImport3));
-        assertThat(thrown.getMessage(), is("Unsupported authorization settings for client 'realm-management' in realm 'realmWithClients': client must be confidential."));
+        assertThat(thrown.getMessage(), is("Unsupported authorization settings for client '${client_realm-management}' in realm 'realmWithClients': client must be confidential."));
 
         doImport("10.4.1_update_realm__raise_error_update_authorization_client_public.json");
         RealmImport foundImport4 = getFirstImport("10.4.2_update_realm__raise_error_update_authorization_client_public.json");
         thrown = assertThrows(ImportProcessingException.class, () -> realmImportService.doImport(foundImport4));
-        assertThat(thrown.getMessage(), is("Unsupported authorization settings for client 'realm-management' in realm 'realmWithClients': client must be confidential."));
+        assertThat(thrown.getMessage(), is("Unsupported authorization settings for client '${client_realm-management}' in realm 'realmWithClients': client must be confidential."));
 
         RealmImport foundImport5 = getFirstImport("10.5_update_realm__raise_error_update_authorization_without_service_account_enabled.json");
         thrown = assertThrows(ImportProcessingException.class, () -> realmImportService.doImport(foundImport5));
-        assertThat(thrown.getMessage(), is("Unsupported authorization settings for client 'realm-management' in realm 'realmWithClients': serviceAccountsEnabled must be 'true'."));
+        assertThat(thrown.getMessage(), is("Unsupported authorization settings for client '${client_realm-management}' in realm 'realmWithClients': serviceAccountsEnabled must be 'true'."));
     }
 
     @Test
@@ -575,7 +582,7 @@ class ImportClientsIT extends AbstractImportTest {
 
         ResourceServerRepresentation authorizationSettings = client.getAuthorizationSettings();
         assertThat(authorizationSettings.getPolicyEnforcementMode(), is(PolicyEnforcementMode.ENFORCING));
-        assertThat(authorizationSettings.isAllowRemoteResourceManagement(), is(false));
+        assertThat(authorizationSettings.isAllowRemoteResourceManagement(), is(true));
         assertThat(authorizationSettings.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
 
         List<ResourceRepresentation> authorizationSettingsResources = authorizationSettings.getResources();
@@ -672,11 +679,54 @@ class ImportClientsIT extends AbstractImportTest {
                 new ScopeRepresentation("urn:servlet-authz:page:main:actionForAdmin"),
                 new ScopeRepresentation("urn:servlet-authz:page:main:actionForUser")
         ));
+
+        client = getClientByName(realm, "missing-id-client");
+        assertThat(client.getName(), is("missing-id-client"));
+        assertThat(client.getClientId(), not(emptyString()));
+        assertThat(client.getDescription(), is("Missing-Id-Client"));
+        assertThat(client.isEnabled(), is(true));
+        assertThat(client.getClientAuthenticatorType(), is("client-secret"));
+        assertThat(client.isServiceAccountsEnabled(), is(true));
+        assertThat(client.getAuthorizationServicesEnabled(), is(true));
+
+        authorizationSettings = client.getAuthorizationSettings();
+        assertThat(authorizationSettings.getPolicyEnforcementMode(), is(PolicyEnforcementMode.ENFORCING));
+        assertThat(authorizationSettings.isAllowRemoteResourceManagement(), is(false));
+        assertThat(authorizationSettings.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
+
+        authorizationSettingsResources = authorizationSettings.getResources();
+        assertThat(authorizationSettingsResources, hasSize(1));
+
+        authorizationSettingsResource = getAuthorizationSettingsResource(authorizationSettingsResources, "Admin Resource");
+        assertThat(authorizationSettingsResource.getUris(), containsInAnyOrder("/protected/admin/*"));
+        assertThat(authorizationSettingsResource.getType(), is("http://servlet-authz/protected/admin"));
+        assertThat(authorizationSettingsResource.getScopes(), containsInAnyOrder(new ScopeRepresentation("urn:servlet-authz:protected:admin:access")));
+
+        authorizationSettingsPolicies = authorizationSettings.getPolicies();
+        authorizationSettingsPolicy = getAuthorizationPolicy(authorizationSettingsPolicies, "Any Admin Policy");
+        assertThat(authorizationSettingsPolicy.getDescription(), is("Defines that adminsitrators can do something"));
+        assertThat(authorizationSettingsPolicy.getType(), is("role"));
+        assertThat(authorizationSettingsPolicy.getLogic(), is(Logic.POSITIVE));
+        assertThat(authorizationSettingsPolicy.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
+        assertThat(authorizationSettingsPolicy.getConfig(), aMapWithSize(1));
+        assertThat(authorizationSettingsPolicy.getConfig(), hasEntry(equalTo("roles"), equalTo("[{\"id\":\"admin\",\"required\":false}]")));
+
+        assertThat(authorizationSettings.getScopes(), hasSize(1));
+        assertThat(authorizationSettings.getScopes(), containsInAnyOrder(
+                new ScopeRepresentation("urn:servlet-authz:protected:admin:access")
+        ));
     }
 
     @Test
     @Order(12)
     void shouldUpdateRealmUpdateAuthorization() throws IOException {
+        // https://github.com/adorsys/keycloak-config-cli/issues/641
+        ResourceRepresentation resource = new ResourceRepresentation();
+        resource.setName("Tweedl Social Service");
+        resource.setType("http://www.example.com/rsrcs/socialstream/140-compatible");
+        resource.setIconUri("http://www.example.com/icons/sharesocial.png");
+        createRemoteManagedClientResource(REALM_NAME, "auth-moped-client", "changed-special-client-secret", resource);
+
         doImport("12_update_realm__update_authorization.json");
 
         RealmRepresentation realm = keycloakProvider.getInstance().realm(REALM_NAME).partialExport(false, true);
@@ -716,7 +766,7 @@ class ImportClientsIT extends AbstractImportTest {
         assertThat(authorizationSettings.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
 
         List<ResourceRepresentation> authorizationSettingsResources = authorizationSettings.getResources();
-        assertThat(authorizationSettingsResources, hasSize(4));
+        assertThat(authorizationSettingsResources, hasSize(5));
 
         ResourceRepresentation authorizationSettingsResource;
         authorizationSettingsResource = getAuthorizationSettingsResource(authorizationSettingsResources, "Admin Resource");
@@ -733,7 +783,6 @@ class ImportClientsIT extends AbstractImportTest {
         assertThat(authorizationSettingsResource.getAttributes(), hasEntry(is("key"), contains("value")));
         assertThat(authorizationSettingsResource.getAttributes(), hasEntry(is("key2"), contains("value2")));
 
-
         authorizationSettingsResource = getAuthorizationSettingsResource(authorizationSettingsResources, "Premium Resource");
         assertThat(authorizationSettingsResource.getUris(), containsInAnyOrder("/protected/premium/*"));
         assertThat(authorizationSettingsResource.getType(), is("urn:servlet-authz:protected:resource"));
@@ -747,6 +796,12 @@ class ImportClientsIT extends AbstractImportTest {
                 new ScopeRepresentation("urn:servlet-authz:page:main:actionForAdmin"),
                 new ScopeRepresentation("urn:servlet-authz:page:main:actionForUser")
         ));
+
+        authorizationSettingsResource = getAuthorizationSettingsResource(authorizationSettingsResources, "Tweedl Social Service");
+        assertThat(authorizationSettingsResource.getUris(), empty());
+        assertThat(authorizationSettingsResource.getType(), is("http://www.example.com/rsrcs/socialstream/140-compatible"));
+        assertThat(authorizationSettingsResource.getIconUri(), is("http://www.example.com/icons/sharesocial.png"));
+        assertThat(authorizationSettingsResource.getScopes(), empty());
 
         List<PolicyRepresentation> authorizationSettingsPolicies = authorizationSettings.getPolicies();
         PolicyRepresentation authorizationSettingsPolicy;
@@ -844,6 +899,42 @@ class ImportClientsIT extends AbstractImportTest {
         assertThat(mopedAuthorizationSettings.getPolicyEnforcementMode(), is(PolicyEnforcementMode.PERMISSIVE));
         assertThat(mopedAuthorizationSettings.isAllowRemoteResourceManagement(), is(true));
         assertThat(mopedAuthorizationSettings.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
+
+        client = getClientByName(realm, "missing-id-client");
+        assertThat(client.getName(), is("missing-id-client"));
+        assertThat(client.getClientId(), not(emptyString()));
+        assertThat(client.getDescription(), is("Missing-Id-Client"));
+        assertThat(client.isEnabled(), is(true));
+        assertThat(client.getClientAuthenticatorType(), is("client-secret"));
+        assertThat(client.isServiceAccountsEnabled(), is(true));
+        assertThat(client.getAuthorizationServicesEnabled(), is(true));
+
+        authorizationSettings = client.getAuthorizationSettings();
+        assertThat(authorizationSettings.getPolicyEnforcementMode(), is(PolicyEnforcementMode.ENFORCING));
+        assertThat(authorizationSettings.isAllowRemoteResourceManagement(), is(true));
+        assertThat(authorizationSettings.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
+
+        authorizationSettingsResources = authorizationSettings.getResources();
+        assertThat(authorizationSettingsResources, hasSize(1));
+
+        authorizationSettingsResource = getAuthorizationSettingsResource(authorizationSettingsResources, "Admin Resource");
+        assertThat(authorizationSettingsResource.getUris(), containsInAnyOrder("/protected/admin/*"));
+        assertThat(authorizationSettingsResource.getType(), is("http://servlet-authz/protected/admin"));
+        assertThat(authorizationSettingsResource.getScopes(), containsInAnyOrder(new ScopeRepresentation("urn:servlet-authz:protected:user:access")));
+
+        authorizationSettingsPolicies = authorizationSettings.getPolicies();
+        authorizationSettingsPolicy = getAuthorizationPolicy(authorizationSettingsPolicies, "Any Admin Policy");
+        assertThat(authorizationSettingsPolicy.getDescription(), is("Defines that adminsitrators can do something"));
+        assertThat(authorizationSettingsPolicy.getType(), is("role"));
+        assertThat(authorizationSettingsPolicy.getLogic(), is(Logic.POSITIVE));
+        assertThat(authorizationSettingsPolicy.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
+        assertThat(authorizationSettingsPolicy.getConfig(), aMapWithSize(1));
+        assertThat(authorizationSettingsPolicy.getConfig(), hasEntry(equalTo("roles"), equalTo("[{\"id\":\"user\",\"required\":false}]")));
+
+        assertThat(authorizationSettings.getScopes(), hasSize(1));
+        assertThat(authorizationSettings.getScopes(), containsInAnyOrder(
+                new ScopeRepresentation("urn:servlet-authz:protected:user:access")
+        ));
     }
 
     @Test
@@ -888,7 +979,7 @@ class ImportClientsIT extends AbstractImportTest {
         assertThat(authorizationSettings.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
 
         List<ResourceRepresentation> authorizationSettingsResources = authorizationSettings.getResources();
-        assertThat(authorizationSettingsResources, hasSize(3));
+        assertThat(authorizationSettingsResources, hasSize(4));
 
         ResourceRepresentation authorizationSettingsResource;
         authorizationSettingsResource = getAuthorizationSettingsResource(authorizationSettingsResources, "Admin Resource");
@@ -908,6 +999,12 @@ class ImportClientsIT extends AbstractImportTest {
                 new ScopeRepresentation("urn:servlet-authz:page:main:actionForPremiumUser"),
                 new ScopeRepresentation("urn:servlet-authz:page:main:actionForAdmin")
         ));
+
+        authorizationSettingsResource = getAuthorizationSettingsResource(authorizationSettingsResources, "Tweedl Social Service");
+        assertThat(authorizationSettingsResource.getUris(), empty());
+        assertThat(authorizationSettingsResource.getType(), is("http://www.example.com/rsrcs/socialstream/140-compatible"));
+        assertThat(authorizationSettingsResource.getIconUri(), is("http://www.example.com/icons/sharesocial.png"));
+        assertThat(authorizationSettingsResource.getScopes(), empty());
 
         List<PolicyRepresentation> authorizationSettingsPolicies = authorizationSettings.getPolicies();
         PolicyRepresentation authorizationSettingsPolicy;
@@ -974,6 +1071,24 @@ class ImportClientsIT extends AbstractImportTest {
         ClientRepresentation mopedClient = getClientByName(realm, "moped-client");
         assertThat(mopedClient.isServiceAccountsEnabled(), is(false));
         assertThat(mopedClient.getAuthorizationSettings(), nullValue());
+
+        client = getClientByName(realm, "missing-id-client");
+        assertThat(client.getName(), is("missing-id-client"));
+        assertThat(client.getClientId(), not(emptyString()));
+        assertThat(client.getDescription(), is("Missing-Id-Client"));
+        assertThat(client.isEnabled(), is(true));
+        assertThat(client.getClientAuthenticatorType(), is("client-secret"));
+        assertThat(client.isServiceAccountsEnabled(), is(true));
+        assertThat(client.getAuthorizationServicesEnabled(), is(true));
+
+        authorizationSettings = client.getAuthorizationSettings();
+        assertThat(authorizationSettings.getPolicyEnforcementMode(), is(PolicyEnforcementMode.ENFORCING));
+        assertThat(authorizationSettings.isAllowRemoteResourceManagement(), is(true));
+        assertThat(authorizationSettings.getDecisionStrategy(), is(DecisionStrategy.UNANIMOUS));
+
+        assertThat(authorizationSettings.getResources(), hasSize(0));
+        assertThat(authorizationSettings.getPolicies(), hasSize(0));
+        assertThat(authorizationSettings.getScopes(), hasSize(0));
     }
 
     @Test
@@ -1685,5 +1800,16 @@ class ImportClientsIT extends AbstractImportTest {
                 .stream()
                 .filter(e -> e.getKey().startsWith(ImportConfigProperties.REALM_STATE_ATTRIBUTE_COMMON_PREFIX))
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+    }
+
+    private void createRemoteManagedClientResource(String realm, String clientId, String clientSecret, ResourceRepresentation resource) {
+        Configuration configuration = new Configuration();
+        configuration.setAuthServerUrl(properties.getUrl().toString());
+        configuration.setRealm(realm);
+        configuration.setResource(clientId);
+        configuration.setCredentials(Collections.singletonMap("secret", clientSecret));
+        AuthzClient authzClient = AuthzClient.create(configuration);
+
+        authzClient.protection().resource().create(resource);
     }
 }